# Reduce compile time of Aggregate