… src/lib/operators/aggregate.hpp