This mostly moves around code in the aggregate operator. It is not the oh-so-much needed cleanup/rewrite, but at least reduces the compile time:

Previously (my laptop, debug): 0m35.766s
Now: 0m22.158s

Details:
* Removes unused GroupByContext struct.
* Splits up AggregateContext so that we can use it without the caller being templated on AggregateKey.
* Stores AggregateResult in a vector instead of the AggregateResultMap. This reduces the number of instantiations of that map. Also, it allows for the output step to be independent of AggregateKey.

Performance, just to show that I didn't break anything (not TPC-H, because that is not using the operator heavily enough; measured on nemea):

```
(release)> generate_tpch 10
(release)> SELECT COUNT(*) FROM lineitem GROUP BY l_suppkey LIMIT 10

# Old: EXECUTE: 5 s 847 ms // 5 s 795 ms // 5 s 733 ms
# New: EXECUTE: 5 s 807 ms // 5 s 699 ms // 5 s 590 ms
```