@mrks 2019-01-16
@lanice: Could you run your benchmarks from the monotonic buffer PR again? The vectors are currently not stored in the buffer, but since they use linear storage, I am not sure if that is necessary.