@lanice 2019-01-16
Hadn't had the time yet to look into this PR more detailed, but here are the benchmark results, current master (including my change) vs this PR:
```
+-----------+------------------+-------+------------------+-------+--------+---------------------------------+
| Benchmark | prev. iter/s     | runs  | new iter/s       | runs  | change | p-value (significant if <0.001) |
+-----------+------------------+-------+------------------+-------+--------+---------------------------------+
| TPC-H 1   | 9.48307132720947 | 569   | 12.2657508850098 | 736   | +29%   |                          0.0000 |
| TPC-H 2   | 5.56658315658569 | 334   | 5.63251066207886 | 338   | +1%    |                          0.4736 |
| TPC-H 3   | 65.5277938842773 | 3932  | 61.6598663330078 | 3700  | -6%    |                          0.0082 |
| TPC-H 4   | 47.7793617248535 | 2867  | 49.247444152832  | 2955  | +3%    |                          0.1700 |
| TPC-H 5   | 62.3465003967285 | 3741  | 63.2133522033691 | 3793  | +1%    |                          0.1997 |
| TPC-H 6   | 440.869354248047 | 10000 | 439.806976318359 | 10001 | -0%    |                          0.0678 |
| TPC-H 7   | 39.580394744873  | 2375  | 40.515022277832  | 2432  | +2%    |                          0.0002 |
| TPC-H 8   | 17.6871318817139 | 1062  | 17.6453151702881 | 1059  | -0%    |                          0.3006 |
| TPC-H 9   | 35.8127250671387 | 2149  | 37.894115447998  | 2274  | +6%    |                          0.0003 |
| TPC-H 10  | 77.2794876098633 | 4637  | 74.8647079467773 | 4492  | -3%    |                          0.0000 |
| TPC-H 11  | 904.392395019531 | 10010 | 952.10791015625  | 10008 | +5%    |                          0.0000 |
| TPC-H 12  | 268.045806884766 | 10000 | 268.1484375      | 10001 | +0%    |                          0.9404 |
| TPC-H 13  | 24.899923324585  | 1494  | 14.749361038208  | 885   | -41%   |                          0.0000 |
| TPC-H 14  | 502.785888671875 | 10001 | 506.671508789062 | 10003 | +1%    |                          0.0188 |
| TPC-H 16  | 51.0632400512695 | 3064  | 52.9981956481934 | 3180  | +4%    |                          0.0000 |
| TPC-H 18  | 32.2972068786621 | 1938  | 22.1995620727539 | 1332  | -31%   |                          0.0000 |
| TPC-H 19  | 319.277770996094 | 10000 | 334.804626464844 | 10005 | +5%    |                          0.0000 |
| TPC-H 22  | 416.774322509766 | 10000 | 412.055450439453 | 10004 | -1%    |                          0.0820 |
| average   |                  |       |                  |       | -1%    |                                 |
+-----------+------------------+-------+------------------+-------+--------+---------------------------------+
```