@lanice 2019-01-17
> Is this correct:

Yes, with scale factor 1 (or is this the default now?). Also, I left out queries 17,20,21, because they are pushing the execution time so much with 100 clients, and I wanted to have the results before the next day ;)

> Also, gcc or clang?

I believe clang