@mrks 2019-01-17
Now:

```
| Benchmark | prev. iter/s  | runs  | new iter/s    | runs  | change | p-value (significant if <0.001) |
+-----------+---------------+-------+---------------+-------+--------+---------------------------------+
| TPC-H 1   | 15.9987716675 | 960   | 12.5492458344 | 753   | -22%   |                          0.0000 |
| TPC-H 2   | 5.89935398102 | 354   | 6.1826672554  | 371   | +5%    |                          0.4477 |
| TPC-H 3   | 68.8749084473 | 4133  | 68.490196228  | 4110  | -1%    |                          0.8288 |
| TPC-H 4   | 45.7643432617 | 2746  | 51.71068573   | 3103  | +13%   |                          0.0000 |
| TPC-H 5   | 59.2104110718 | 3553  | 65.9138717651 | 3955  | +11%   |                          0.0000 |
| TPC-H 6   | 453.210083008 | 10000 | 463.758575439 | 10001 | +2%    |                          0.0000 |
| TPC-H 7   | 31.8477039337 | 1911  | 41.2496604919 | 2475  | +30%   |                          0.0000 |
| TPC-H 8   | 18.911151886  | 1135  | 19.2568340302 | 1156  | +2%    |                          0.2828 |
| TPC-H 9   | 35.4976882935 | 2130  | 39.2272758484 | 2354  | +11%   |                          0.0000 |
| TPC-H 10  | 82.7401046753 | 4965  | 82.8384246826 | 4971  | +0%    |                          0.1309 |
| TPC-H 11  | 990.130004883 | 10006 | 954.718688965 | 10007 | -4%    |                          0.0000 |
| TPC-H 12  | 281.468048096 | 10001 | 297.35333252  | 10000 | +6%    |                          0.0000 |
| TPC-H 13  | 18.6318798065 | 1118  | 14.6310377121 | 878   | -21%   |                          0.0000 |
| TPC-H 14  | 545.170959473 | 10003 | 565.148132324 | 10005 | +4%    |                          0.0000 |
| TPC-H 16  | 34.6303443909 | 2078  | 59.2283859253 | 3554  | +71%   |                          0.0000 |
| TPC-H 18  | 30.863368988  | 1852  | 34.4618606567 | 2068  | +12%   |                          0.0000 |
| TPC-H 19  | 343.730651855 | 10000 | 348.419769287 | 10000 | +1%    |                          0.0305 |
| TPC-H 22  | 416.337127686 | 10008 | 422.568634033 | 10007 | +1%    |                          0.0543 |
| average   |               |       |               |       | +7%    |                                 |
+-----------+---------------+-------+---------------+-------+--------+---------------------------------+
```