@Bouncner 2019-01-18
Since 1 and 13 are "slower" than 16 (which gained a lot) already: any ideas why we lose performance here?