@mrks 2019-01-18
Instead of storing the result struct in the map, we have an additional indirection into a vector. That seems to be good for many queries, but not so good for others. I would agree with @lanice and accept those changes for now.