@mrks 2019-01-18
Might also be random variations, because TPC-H 1 was not as bad in the first benchmark.